requirements.txt
pytorch==1.0
numpy==1.13.1
tqdm==4.15.0
torchtext==0.3.1
nltk==3.4
scikit-learn==0.20.3
bpemb==0.3.0
scipy==1.2.1
